source_backup/requirements.txt
aiofiles==23.2.1
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anthropic==0.33.1
anyio==4.4.0
async-timeout==4.0.3
attrs==23.2.0
cachetools==4.2.4
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.1
cryptography==43.0.0
cycler==0.12.1
defusedxml==0.7.1
distro==1.9.0
dnspython==2.6.1
docx==0.2.4
email_validator==2.2.0
et-xmlfile==1.1.0
exceptiongroup==1.2.2
fastapi==0.111.1
fastapi-cli==0.0.4
ffmpy==0.4.0
filelock==3.15.4
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.6.1
google-api-core==1.34.1
google-auth==1.35.0
google-cloud-core==1.7.3
google-cloud-translate==2.0.1
googleapis-common-protos==1.63.2
gradio==4.40.0
gradio_client==1.2.0
grpc-google-iam-v1==0.13.1
grpcio==1.65.4
grpcio-status==1.48.2
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.24.3
idna==3.7
importlib_resources==6.4.0
Jinja2==3.1.4
jiter==0.5.0
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.5
langchain==0.2.11
langchain-anthropic==0.1.22
langchain-core==0.2.30
langchain-openai==0.1.17
langchain-text-splitters==0.2.2
langsmith==0.1.93
lxml==5.2.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.1
mdurl==0.1.2
multidict==6.0.5
numpy==1.26.4
openai==1.37.0
openpyxl==3.1.5
orjson==3.10.6
packaging==24.1
pandas==2.2.2
pdfminer.six==20231228
pdfplumber==0.11.3
pillow==10.4.0
proto-plus==1.24.0
protobuf==3.20.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.8.2
pydantic-settings==2.3.4
pydantic_core==2.20.1
pydub==0.25.1
Pygments==2.18.0
pyparsing==3.1.2
PyPDF2==3.0.1
pypdfium2==4.30.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-multipart==0.0.9
python-pptx==0.6.23
pytz==2024.1
PyYAML==6.0.1
regex==2024.7.24
requests==2.32.3
rich==13.7.1
rsa==4.9
ruff==0.5.5
semantic-version==2.10.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.31
starlette==0.37.2
tenacity==8.5.0
tiktoken==0.7.0
tokenizers==0.20.0
tomlkit==0.12.0
tqdm==4.66.4
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
uvicorn==0.30.3
uvloop==0.19.0
watchfiles==0.22.0
websockets==11.0.3
XlsxWriter==3.2.0
yarl==1.9.4
